+++ src/components/modules/modificationsObserver.ts
@@ -32,7 +32,6 @@ export default class ModificationsObserver extends Module {
    * @type {Function}
    */
   private mutationDebouncer = _.debounce( () => {
-    this.checkEmptiness();
     this.config.onChange();
   }, ModificationsObserver.DebounceTimer);
 
@@ -142,14 +141,5 @@ export default class ModificationsObserver extends Module {
     if (contentMutated) {
       this.mutationDebouncer();
     }
-  }
-
-  /**
-   * Check if Editor is empty and set CSS class to wrapper
-   */
-  private checkEmptiness(): void {
-    const {BlockManager, UI} = this.Editor;
-
-    UI.nodes.wrapper.classList.toggle(UI.CSS.editorEmpty, BlockManager.isEditorEmpty);
   }
 }